To do list was created.
To do lists were coded to be shown according to dates.
To do list connected to cloud firestore.
Reminder will be fixed.
Recurring UI will be fixed.
To do lists will be set to be user specific.
To do lists to be made will be organized to be attached to the dates